Reorganizes tests into a descriptive namespace.
