+++ tests/test_primitives/test_point/test_point_serializers.py
